Refactor unsigned long ref #35

+++ solveQuery.h
@@ -2,6 +2,6 @@
     #define SOLVE_QUERY_H
     #include "utility.h"
 
-    bool dfs_search(row_g *graph, int node1, int node2, bool *visited);
+    bool dfs_search(row_g *graph, unsigned long node1, unsigned long node2, bool *visited);
     void *solveQuery (void *args);
 #endif  //SOLVE_QUERY_H
